+++ src/main/kotlin/com/app/drivn/backend/user/dto/UserInfoDto.kt
@@ -12,5 +12,5 @@ open class UserInfoDto {
   lateinit var tokenClaimable: BigDecimal
   lateinit var tokensLimitPerDay: BigDecimal
   var nextEnergyRenew: ZonedDateTime? = null
-  var donation by Delegates.notNull<Int>()
+  var donation by Delegates.notNull<Short>()
 }
